requirements.txt
alembic==1.15.2
cryptography==44.0.3
email-validator==2.2.0
email_validator==2.2.0
flask==3.1.1
flask-dance==7.1.0
flask-login==0.6.3
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
Jinja2==3.1.6
oauthlib==3.2.2
pdfminer.six==20250506
python==3.11.12
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
sqlalchemy==2.0.40
werkzeug==3.1.3
wtforms==3.2.1
